create CUDAExt
