+++ Project.toml
@@ -1,7 +1,7 @@
 name = "ObjectDetector"
 uuid = "3dfc1049-5314-49cf-8447-288dfd02f9fb"
 authors = ["Robert Luciani"]
-version = "0.2.10"
+version = "0.3.0"
 
 [deps]
 BenchmarkTools = "6e4b80f9-dd63-53aa-95a3-0cdb28fa8baf"
@@ -17,6 +17,13 @@ PrecompileTools = "aea7be01-6a6a-4083-8856-8a6e6704d82a"
 PrettyTables = "08abe8d2-0d0c-5749-adfa-8a2ac140af0d"
 cuDNN = "02a925ec-e4fe-4b08-9a7e-0d78e3d38ccd"
 
+[weakdeps]
+CUDA = "052768ef-5323-5732-b1bb-66c8b64840ba"
+cuDNN = "02a925ec-e4fe-4b08-9a7e-0d78e3d38ccd"
+
+[extensions]
+CUDAExt = ["CUDA", "cuDNN"]
+
 [compat]
 BenchmarkTools = "0.4, 0.5, 0.6, 0.7, 1.0"
 CUDA = "4"
@@ -26,10 +33,9 @@ ImageDraw = "0.2"
 ImageFiltering = "0.6, 0.7"
 ImageTransformations = "0.8, 0.9"
 LazyArtifacts = "1.3"
-PrettyTables = "2.0"
 PrecompileTools = "1"
-cuDNN = "1"
-julia = "1.3"
+PrettyTables = "2.0"
+julia = "1.6"
 
 [extras]
 FileIO = "5789e2e9-d7fb-5bc7-8068-2c6fae9b9549"
